Move all `ReactSpan` subclasses to a separate folder
