+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/text/ReactBaseTextShadowNode.java
@@ -31,6 +31,21 @@ import com.facebook.react.uimanager.ReactAccessibilityDelegate.Role;
 import com.facebook.react.uimanager.ReactShadowNode;
 import com.facebook.react.uimanager.ViewProps;
 import com.facebook.react.uimanager.annotations.ReactProp;
+import com.facebook.react.views.text.span.CustomLetterSpacingSpan;
+import com.facebook.react.views.text.span.CustomLineHeightSpan;
+import com.facebook.react.views.text.span.CustomStyleSpan;
+import com.facebook.react.views.text.span.ReactAbsoluteSizeSpan;
+import com.facebook.react.views.text.span.ReactBackgroundColorSpan;
+import com.facebook.react.views.text.span.ReactClickableSpan;
+import com.facebook.react.views.text.span.ReactForegroundColorSpan;
+import com.facebook.react.views.text.span.ReactSpan;
+import com.facebook.react.views.text.span.ReactStrikethroughSpan;
+import com.facebook.react.views.text.span.ReactTagSpan;
+import com.facebook.react.views.text.span.ReactUnderlineSpan;
+import com.facebook.react.views.text.span.SetSpanOperation;
+import com.facebook.react.views.text.span.ShadowStyleSpan;
+import com.facebook.react.views.text.span.TextInlineImageSpan;
+import com.facebook.react.views.text.span.TextInlineViewPlaceholderSpan;
 import com.facebook.yoga.YogaDirection;
 import com.facebook.yoga.YogaUnit;
 import com.facebook.yoga.YogaValue;
@@ -366,15 +381,16 @@ public abstract class ReactBaseTextShadowNode extends LayoutShadowNode
     // or images.
     for (int priorityIndex = 0; priorityIndex < ops.size(); priorityIndex++) {
       final SetSpanOperation op = ops.get(ops.size() - priorityIndex - 1);
+      final ReactSpan what = op.getWhat();
 
-      boolean isInlineImage = op.what instanceof TextInlineImageSpan;
-      if (isInlineImage || op.what instanceof TextInlineViewPlaceholderSpan) {
+      boolean isInlineImage = what instanceof TextInlineImageSpan;
+      if (isInlineImage || what instanceof TextInlineViewPlaceholderSpan) {
         int height;
         if (isInlineImage) {
-          height = ((TextInlineImageSpan) op.what).getHeight();
+          height = ((TextInlineImageSpan) what).getHeight();
           textShadowNode.mContainsImages = true;
         } else {
-          TextInlineViewPlaceholderSpan placeholder = (TextInlineViewPlaceholderSpan) op.what;
+          TextInlineViewPlaceholderSpan placeholder = (TextInlineViewPlaceholderSpan) what;
           height = placeholder.getHeight();
 
           // Inline views cannot be layout-only because the ReactTextView needs to be able to grab
