+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/text/span/ReactTagSpan.java
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package com.facebook.react.views.text;
+package com.facebook.react.views.text.span;
 
 /**
  * Instances of this class are used to place reactTag information of nested text react nodes into
