+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/text/span/SetSpanOperation.java
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package com.facebook.react.views.text;
+package com.facebook.react.views.text.span;
 
 import android.text.Spannable;
 import android.text.SpannableStringBuilder;
@@ -17,9 +17,10 @@ public class SetSpanOperation {
   static final int SPAN_MAX_PRIORITY = Spanned.SPAN_PRIORITY >> Spanned.SPAN_PRIORITY_SHIFT;
 
   protected int start, end;
-  protected ReactSpan what;
 
-  SetSpanOperation(int start, int end, ReactSpan what) {
+  private final ReactSpan what;
+
+  public SetSpanOperation(int start, int end, ReactSpan what) {
     this.start = start;
     this.end = end;
     this.what = what;
@@ -55,5 +56,9 @@ public class SetSpanOperation {
     spanFlags |= (effectivePriority << Spannable.SPAN_PRIORITY_SHIFT) & Spannable.SPAN_PRIORITY;
 
     sb.setSpan(what, start, end, spanFlags);
+  }
+
+  public ReactSpan getWhat() {
+    return what;
   }
 }
