+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/text/span/TextInlineImageSpan.java
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package com.facebook.react.views.text;
+package com.facebook.react.views.text.span;
 
 import android.graphics.drawable.Drawable;
 import android.text.Spannable;
