+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/textinput/ReactEditText.java
@@ -52,19 +52,19 @@ import com.facebook.react.uimanager.ReactAccessibilityDelegate;
 import com.facebook.react.uimanager.StateWrapper;
 import com.facebook.react.uimanager.UIManagerModule;
 import com.facebook.react.uimanager.events.EventDispatcher;
-import com.facebook.react.views.text.CustomLetterSpacingSpan;
-import com.facebook.react.views.text.CustomLineHeightSpan;
-import com.facebook.react.views.text.CustomStyleSpan;
-import com.facebook.react.views.text.ReactAbsoluteSizeSpan;
-import com.facebook.react.views.text.ReactBackgroundColorSpan;
-import com.facebook.react.views.text.ReactForegroundColorSpan;
-import com.facebook.react.views.text.ReactSpan;
-import com.facebook.react.views.text.ReactStrikethroughSpan;
+import com.facebook.react.views.text.span.CustomLetterSpacingSpan;
+import com.facebook.react.views.text.span.CustomLineHeightSpan;
+import com.facebook.react.views.text.span.CustomStyleSpan;
+import com.facebook.react.views.text.span.ReactAbsoluteSizeSpan;
+import com.facebook.react.views.text.span.ReactBackgroundColorSpan;
+import com.facebook.react.views.text.span.ReactForegroundColorSpan;
+import com.facebook.react.views.text.span.ReactSpan;
+import com.facebook.react.views.text.span.ReactStrikethroughSpan;
 import com.facebook.react.views.text.ReactTextUpdate;
 import com.facebook.react.views.text.ReactTypefaceUtils;
-import com.facebook.react.views.text.ReactUnderlineSpan;
+import com.facebook.react.views.text.span.ReactUnderlineSpan;
 import com.facebook.react.views.text.TextAttributes;
-import com.facebook.react.views.text.TextInlineImageSpan;
+import com.facebook.react.views.text.span.TextInlineImageSpan;
 import com.facebook.react.views.text.TextLayoutManager;
 import com.facebook.react.views.view.ReactViewBackgroundManager;
 import java.util.ArrayList;
